[Not for upstream] add header file with GPIO macros
This commit introduces the gpio-starfive-vic.h header file separate from
the gpio driver as it is not actually used by the gpio driver anymore.

This file contains macros that substitute for proper gpio and pinctrl
kernel code. The macros are used in fuction i2c_dw_configure_gpio()
in drivers/i2c/busses/i2c-designware-master.c. This can be carried as a
downstream patch until the hacky code in i2c_dw_configure_gpio() is
rewritten following the creation of a pinctrl driver.

Signed-off-by: Drew Fustini <drew@beagleboard.org>