pinctrl: starfive: improve register field documentation
This commit should eventually be squashed.

Add information about the register offsets and field descriptions from
the JH7100 datasheet in particular regarding the registers related to
GPIO interrupt. I found the naming used in the original starfive code to
be confusing and I had to keep flipping back and forth between the
datasheet. I felt that this would make it easier for people to read and
review the code if there was more inline documentation. I also renamed
some of the defines and variables to make the intent more clear.

NOTE: Table 12-9 in the JH7100 datasheet is incorrect regarding fields
GPIOIE_0 and GPIOIE_1. An interrupt is enabled (unmasked) when the bit
is set to 1 and it is disabled (masked) when set to 0. The datasheet
incorrectly states the opposite. I think this is due to the datasheet
author thinking of it as mask field which it is not, it is an enable
field. I will raise an issue on the documentation repo.

Signed-off-by: Drew Fustini <drew@pdp7.com>