clk: starfive: jh7100: Don't round divider up twice
The problem is best illustrated by an example. Suppose a consumer wants
a 7MHz clock rate from a divider with a 100MHz parent. It would then
call

  clk_round_rate(clk, 7000000)

which would call into our determine_rate() callback that would correctly
determine that a divider of 15 gives the highest frequency below the
requested 7MHz and return 100000000 / 15 = 6666666Hz.

However the consumer would then call

  clk_set_rate(clk, 6666666)

but since 6666666 doesn't divide 100000000 evenly our set_rate()
callback would again round the divider up and set it to 16 which results
in an unnecessarily low rate of 6.25MHz.

Fix it by using DIV_ROUND_CLOSEST in the set_rate() callback.

Signed-off-by: Emil Renner Berthing <kernel@esmil.dk>