nvmem: core: Support reading cells with >= 8 bit offsets
For NVMEM devices with .stride > 1, some cell values may not be aligned
to the device's stride. In this case, it is necessary to use bit_offset
to access the cell. For example, to access the third byte of an NVMEM
device with .stride == 4, we need "bits = <16 8>;" in the devicetree.

Implement this on the read side. The write side implementation would be
more complicated, and it is not necessary for read-only NVMEM devices.
For now, reject writes for these cells to avoid any incorrect behavior.

Signed-off-by: Samuel Holland <samuel@sholland.org>
(forward ported from commit b24df690d77a6720cd3cb515590460a097de42eb https://github.com/smaeul/linux.git)
[esmil: rename variables to match struct field rename, remove info.bit_offset >= BITS_PER_BYTE check]
Signed-off-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>