riscv: errata: Add EIC7700 non-coherent DMAs
Similar to the StarFive JH7100 the ESWIN EIC7700 SoC does not implement
the standard Zicbom and Svpbm extensions but instead relies on the
SiFive cache controller to do cache flushes and maps the external memory
multiple times in the physical address space - cached in one place and
uncached at a higher address. To handle this we need the DMA_GLOBAL_POOL
and RISCV_NONSTANDARD_CACHE_OPS enabled to work correctly.

Signed-off-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>