pinctrl: spacemit: introduce syscon and regmap
On wakeup from suspend, only pinctrl-level interrupts
are triggered, while GPIO-level interrupts are not.

However, pinctrl registers do not provide sufficient
information to determine the wake source or edge polarity.

Retrieving this information requires access to registers
in the GPIO module.

To safely share access to these registers with the
GPIO driver, we introduce a syscon binding and use
regmap for coordinated access.

Change-Id: Iecf7aec836e77c6200b597616154b65e2d0e844f
Signed-off-by: Zexi Guo <zexi.guo@spacemit.com>