riscv: dts: k3: fix syscon(gpio & pinctrl using) phandle naming
The pinctrl node was using properties named "syscon-gpio" and
"syscon-gpio-edge".

Since the property names contain "-gpio", the OF core mistakenly
interprets these phandles as GPIO specifiers and attempts to
parse #gpio-cells, which are not defined for the syscon nodes.
This leads to warnings like:

  OF: /soc/gpio@d4019000: could not get #gpio-cells for /soc/syscon-gpio
  OF: /soc/pinctrl@d401e000: could not get #gpio-cells for /soc/syscon-gpio

Rename the properties to "syscon" and "syscon-edge" to avoid the
parsing issue while still keeping the syscon nodes functional.

Change-Id: I86dd39593de454ef723e4064e240b9b5c13f8ec7
Reported-by: Juan Li <juan.li@spacemit.com>
Signed-off-by: Zexi Guo <zexi.guo@spacemit.com>