pinctrl: spacemit: k1: fix syscon phandle naming to match DTS
The DTS for syscon nodes were recently updated to rename
the properties "syscon-gpio" and "syscon-gpio-edge" to "syscon"
and "syscon-edge".

The old names caused the OF core to misinterpret the phandles as
GPIO specifiers and resulted in #gpio-cells warnings.

Update the driver accordingly to use the new property names,
so that it can properly lookup the regmaps without errors.

Change-Id: Ia7a390d7fb1eb1ba2306b3ec09a71567ddbc4d16
Reported-by: Juan Li <juan.li@spacemit.com>
Signed-off-by: Zexi Guo <zexi.guo@spacemit.com>