usb: dwc3: generic: add support for spacemit,k1-dwc3
SpacemiT DWC3 controller has vendor registers to support USB wakeup,
support it according to compatible match.

If reset is managed by DWC3 core, it will reset controller when USB
wakeup is not enabled. Add a reset-on-resume property to control reset
behavior of generic-plat driver during suspend, and it could be passed
to xhci driver.

Change-Id: Ia2c9b74ced9688181a5fa6e1910d66c83cf2accb
Signed-off-by: Junzhong Pan <junzhong.pan@spacemit.com>