perf diff: Constify strchr() return variables
Newer glibc versions return const char for strchr() when the 's' arg is
const, change the return variable to const to match that.

Also we don't need to turn that ',' into a '\0', as strtol() will stop
in the first invalid char. No need to touch read only memory.

First noticed with fedora 44.

Reviewed-by: Ian Rogers <irogers@google.com>
Link: https://lore.kernel.org/r/20251211221756.96294-3-acme@kernel.org
Change-Id: I3932b4e6ffa46966b8486fb75c3bcf6b06406cdb
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>