spacemit: usb: dwc3: gadget: notify disconnect from role-switch none
The disconnect event of dwc3 gadget rely on vbus detection. For typec
connector, the attach/detach is detected by typec chip.

If there is no disconnect event, the usb state cannot restore for
usb2.0 re-enumeration after a USB3 unplug, USB would not be recognized.

Add support to let role-switch to notify dwc3 that a disconnection
happends.

Change-Id: I1de6b550a45164828c8c299123a59cc42e7ab2c8
Signed-off-by: Junzhong Pan <junzhong.pan@spacemit.com>