pinctrl: spacemit: change strict to false
When the strict member is set to true, a GPIO cannot be requested while
simultaneously configuring pull-up/down and drive strength.

This patch works around the issue by setting strict to false. Note that
this is very likely a workaround; the patch will be reverted once a
proper solution is available.

Change-Id: I7b1cda00c6d599caa139e519b90e12f0a671ca99
Signed-off-by: Zexi Guo <zexi.guo@spacemit.com>